requirements.txt
python-dotenv>=1.0.0
langchain>=1.2.0
langgraph>=0.1.0
langchain-openai>=0.0.8
openai>=0.27.0
requests>=2.28.0